Add extractor to process the file retireved from the MTA
The status file is an XML file. The status_extractor process the XML to
extract the subway lines. The extractor currently handles lines that do
not exist and lines that are not grouped (eg. 7 or G, but not ACE).

The extractor currently returns '*GOOD SERVICE*' for all valid non
grouped lines.