Status extracter now handles lines that have 'DELAY' as their status
The extractor now processes the text file response that comes back from
the MTA website. It gives back the status, the date and time the status
was posted and any message that goes with the status of the line.

Cleanup of the status message is needed due to HTML tags etc.