Status extractor generates a custom message for lines not directly affected by status change
Lines may have a status change due to another line. Instead of
presenting the status message that is related to those other lines, the
status extractor provides a custom message indicating that the line is
not directly affected and which lines are directly affected which can
then be queried to get the original status message.